requirements.txt
aiosmtpd==1.4.4.post2
alembic==1.13.1
APScheduler==3.10.4
atpublic==4.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.5.0
filelock==3.13.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-msearch==0.2.9.4
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.49.0
fsspec==2024.2.0
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.0
pillow==10.2.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.16
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.2
